requirements.txt
apache-airflow-providers-google >= 6.2.0
google-cloud-bigquery >= 2.31.0
google-cloud-bigquery-datatransfer >= 3.4.1
google-cloud-bigquery-storage >= 2.10.1
google-api-core >= 1.31.4
airflow-provider-great-expectations >= 0.2.2
pyarrow >= 5.0.0
pandas >= 1.3.5
great_expectations >= 0.15.41
openpyxl >= 3.0.10
sqlalchemy-bigquery >= 1.5.0
Flask >= 1.1.2
SQLAlchemy >= 1.3.24
croniter >= 1.0.15